# docs/api.yml
openapi: 3.0.3
info:
  title: Docbox API
  version: 1.0.0
  description: API specification for docbox
paths:
...
  /box/{scope}/file/presigned:
    parameters:
      - $ref: "#/components/parameters/x-tenant-env"
      - $ref: "#/components/parameters/x-tenant-id"
      - $ref: "#/components/parameters/x-user-id"
      - $ref: "#/components/parameters/x-user-name"
      - $ref: "#/components/parameters/x-user-image-id"
      - $ref: "#/components/parameters/scope"
    post:
      summary: Create a presigned upload
      operationId: create_presigned
      description: >
        Creates a presigned upload task to allow uploading a file directly to S3
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                folder_id:
                  allOf:
                    - $ref: "#/components/schemas/FolderId"
                    - description: ID of the folder to store the file in
                size:
                  type: number
                  description: Size of the file that will be uploaded
                mime:
                  type: string
                  description: Mime/Content type of the file that will be uploaded

      responses:
        "201":
          description: Created a presigned upload
          content:
            application/json:
              schema:
                type: object
                properties:
                  task_id:
                    allOf:
                      - $ref: "#/components/schemas/PresignedUploadTaskId"
                        description: ID of the presigned upload task to poll for completion
                  method:
                    type: string
                    description: HTTP method to perform the presigned upload request with
                  uri:
                    type: string
                    description: URL to send the presigned upload request to
                  headers:
                    type: object
                    description: Any headers that must be included in the request
                required:
                  - file
                  - name
                  - folder_id
  /box/{scope}/file/presigned/{task_id}:
    parameters:
      - $ref: "#/components/parameters/x-tenant-env"
      - $ref: "#/components/parameters/x-tenant-id"
      - $ref: "#/components/parameters/x-user-id"
      - $ref: "#/components/parameters/x-user-name"
      - $ref: "#/components/parameters/x-user-image-id"
      - $ref: "#/components/parameters/scope"
      - $ref: "#/components/parameters/presigned_upload_task_id"
    get:
      summary: Get a presigned upload
      operationId: get_presigned
      description: >
        Get the status of a presigned file upload
      responses:
        "200":
          description: Created a presigned upload
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      status:
                        type: string
                        enum: [Pending]
                  - type: object
                    properties:
                      status:
                        type: string
                        enum: [Complete]
                      file:
                        $ref: "#/components/schemas/FileWithExtra"
                      generated:
                        type: array
                        items:
                          $ref: "#/components/schemas/GeneratedFile"
                  - type: object
                    properties:
                      status:
                        type: string
                        enum: [Failed]
                      error:
                        type: string
        "404":
          description: Presigned upload task not found
        "500":
          $ref: "#/components/responses/InternalServerError"
  /box/{scope}/file/{file_id}:
    parameters:
      - $ref: "#/components/parameters/x-tenant-env"
      - $ref: "#/components/parameters/x-tenant-id"
      - $ref: "#/components/parameters/x-user-id"
      - $ref: "#/components/parameters/x-user-name"
      - $ref: "#/components/parameters/x-user-image-id"
      - $ref: "#/components/parameters/scope"
      - $ref: "#/components/parameters/file_id"
    get:
      summary: Get a file
      operationId: get_file
      description: >
        Request details about a specific file, can be used to check if a file exists.

        Response also includes a list of generated files which can be used to 
        determine if the file can be previewed in a specific way
      responses:
        "200":
          description: File existed
          content:
            application/json:
              schema:
                type: object
                properties:
                  file:
                    $ref: "#/components/schemas/FileWithExtra"
                  generated:
                    $ref: "#/components/schemas/GeneratedFile"
        "404":
          description: File not found
        "500":
          $ref: "#/components/responses/InternalServerError"
    put:
      summary: Update a file
      operationId: update_file
      description: >
        Update a file, this can be used to change the file name and move 
        the file to another folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The new name for the file if updating the file
                  minLength: 1
                  nullable: true
                folder_id:
                  allOf:
                    - $ref: "#/components/schemas/FolderId"
                    - nullable: true
                      description: New folder to move to if moving the file
      responses:
        "200":
          description: File updated successfully
        "404":
          description: File not found
        "500":
          $ref: "#/components/responses/InternalServerError"
    delete:
      summary: Delete a file
      operationId: delete_file
      description: >
        Deletes the file and all associated generated files
      responses:
        "204":
          description: File deleted successfully
        "404":
          description: File not found
        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
  parameters:
    scope:
      name: scope
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/DocumentBoxScope"
...
    file_id:
      name: file_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/FileId"
...
    presigned_upload_task_id:
      name: task_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/PresignedUploadTaskId"
...
    x-tenant-id:
      name: x-tenant-id
      in: header
      description: ID of the tenant the request is for
      required: true
      schema:
        $ref: "#/components/schemas/TenantId"
    x-tenant-env:
      name: x-tenant-env
      in: header
      description: Environment to operate withins
      required: true
      schema:
        type: string
    x-user-id:
      name: x-user-id
      in: header
      description: ID of the user performing the operation (If performed by a user)
      required: false
      schema:
        $ref: "#/components/schemas/UserId"
    x-user-name:
      name: x-user-id
      in: header
      description: Name of the user performing the operation (If performed by a user)
      required: false
      schema:
        type: string
    x-user-image-id:
      name: x-user-image-id
      in: header
      description: Image ID of the user performing the operation (If performed by a user)
      required: false
      schema:
        type: string
  responses:
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    UserId:
      type: string
      description: ID of a user
    TenantId:
      type: string
      format: uuid
      description: ID of a tenant
...
    PresignedUploadTaskId:
      type: string
      format: uuid
      description: ID of a presigned upload task that can be polled
    FileId:
      type: string
      format: uuid
      description: ID of a file
    FolderId:
      type: string
      format: uuid
      description: ID of a folder
...
    DocumentBoxScope:
      type: string
      description: Unique scope for a document box
      minLength: 1
      example: "user:a71502fe-c025-4a06-ae5d-3131535a8237:uploaded"
...
    FileWithExtra:
      type: object
      description: >
        File with additional information about the most recent edit and user information
      properties:
        id:
          $ref: "#/components/schemas/FileId"
        name:
          type: string
          description: Unique ID of the file
        mime:
          type: string
          description: MIME / content type of the file
        folder_id:
          allOf:
            - $ref: "#/components/schemas/FolderId"
            - description: ID of the folder the file is within
        hash:
          type: string
          description: SHA256 hash of the file contents
        size:
          type: number
          description: File size in bytes
        encrypted:
          type: boolean
          description: Whether the file has been detected as encrypted during processing
        created_at:
          type: string
          format: date-time
          description: When the file was created
        created_by:
          allOf:
            - $ref: "#/components/schemas/User"
            - nullable: true
              description: User who created the file (If created by a user)
        last_modified_at:
          type: string
          format: date-time
          nullable: true
          description: The last datetime of modification if the file has been modified
        last_modified_by:
          allOf:
            - $ref: "#/components/schemas/User"
            - nullable: true
              description: The last user who performed the last modification (If available)
      required:
        - id
        - name
        - mime
        - folder_id
        - hash
        - encrypted
        - created_at
...
    GeneratedFileType:
      type: string
      description: Type of the generated file (PDF, thumbnail etc)
      enum: [Pdf, CoverPage, SmallThumbnail, LargeThumbnail, TextContent]
    GeneratedFile:
      type: object
      description: >
        File generated as the outcome of processing another image file, could be a converted
        file or a thumbnail image depending on the type
      properties:
        id:
          type: string
        file_id:
          allOf:
            - $ref: "#/components/schemas/FileId"
            - description: ID of the file that this was generated from
        mime:
          type: string
          description: Mime type of the generated file
        type:
          $ref: "#/components/schemas/GeneratedFileType"
        hash:
          type: string
          description: SHA256 hash of the generated file contents
        created_at:
          type: string
          format: date-time
          description: When the file was created
      required:
        - id
        - file_id
        - mime
        - type
        - hash
        - created_at
    User:
      type: object
      description: >
        Snapshot of the external user who performed the action,
        the details for the user are updated every time they
        perform an action
      properties:
        id:
          allOf:
            - $ref: "#/components/schemas/UserId"
            - type: string
              description: External ID of the user
        name:
          type: string
          nullable: true
          description: External name of the user if known
        image_id:
          type: string
          nullable: true
          description: External image ID of the user if known
      required:
        - id
...
    ErrorResponse:
      type: object
      properties:
        reason:
          type: string
          description: Descriptive message for the cause of the error
        backtrace:
          type: string
          description: >
            Server backtrace of the error, only present in local debugging disabled
            for production releases